+++ src/store/team/types.js
@@ -1,0 +1,7 @@
+// Team types
+
+// Successfully loaded team member info
+export const LOAD_TEAM_MEMBER_SUCCESS = 'LOAD_TEAM_MEMBER_SUCCESS';
+
+// Successfully loaded team members info
+export const LOAD_TEAM_MEMBERS_SUCCESS = 'LOAD_TEAM_MEMBERS_SUCCESS';